Adds API routes for creating and viewing orders and removing order li…

+++ app/app.py
@@ -3,7 +3,7 @@ import datetime
 from flask import Flask, jsonify, render_template, request, redirect, url_for
 from flask_login import LoginManager, login_user, logout_user, login_required, UserMixin, current_user
 from database.database import db
-from database.models import User, Product, Cache, Product
+from database.models import User, Product, Cache, Product, Order, OrderLine
 
 app = Flask(__name__, static_url_path='/static')
 app.config['SECRET_KEY'] = 'secret'
@@ -74,8 +74,8 @@ def create_account() -> str:
 def store() -> str:
     return render_template("storeFront.html")
 
-@app.route("/api/product/<int:product_id>", methods=["GET"])
-def retrive_product(product_id):
+@app.route("/store/<int:product_id>", methods=["GET"])
+def retreive_product(product_id):
     try:
         prod = db.session.get(Product, product_id)
         return (
@@ -88,9 +88,11 @@ def retrive_product(product_id):
             404
         )
 
-@app.route("/store/<int:product_id>", methods=["DELETE"])
+@app.route("/admin/store/<int:product_id>", methods=["DELETE"])
 def remove_item(product_id):
     try:
+        if current_user.__auth__() != True:
+                raise PermissionError
         product = db.session.get(Product, product_id)
         db.session.delete(product)
         db.session.commit()
@@ -98,64 +100,98 @@ def remove_item(product_id):
             f"Product #{product_id} removed successfully.",
             200
         )
+    except PermissionError:
+        return (
+            f"err",
+            401
+        )
     except:
         return (
             f"Product #{product_id} could not be found.",
             404
         )
 
-@app.route("/store/newproduct", methods=["POST"])
+@app.route("/admin/store/newproduct", methods=["POST"])
 def create_product():
+    data = request.json
     try:
-        data = request.json
-        new_prod = Product(
-            name=data.get("name"),
-            price=data.get("price"),
-            quantity=data.get("quantity")
-            )
-        db.session.add(new_prod)
-        db.session.commit()
+            if current_user.__auth__() != True:
+                raise PermissionError
+            
+            err_list = []
+            err_string = ""
+            
+            if type(data["name"]) != type("e"):
+                err_list.append("name must be string")
+            if type(data["price"]) != type(0) and type(data["price"]) != type(0.0):
+                err_list.append("price must be numeric")
+            if type(data["quantity"]) != type(0):
+                err_list.append("quantity must be an integer")
+            
+            if len(err_list) > 0:
+                for item in err_list:
+                    err_string += f"{item}"
+                    if item != err_list[-1]:
+                        err_string += ", "
+                raise AttributeError
+            
+    except AttributeError:
         return (
-            f"Product #{new_prod.id} added successfully.",
-            200
+            f"Invalid format for product attributes: {err_string}.",
+            400
+        )
+    
+    except PermissionError:
+        return (
+            f"err", #use permissions code
+            401
         )
         
-    except:
-        return (
-            "err num",
-            "err msg"
+    new_prod = Product(
+        name=data.get("name"),
+        price=round(float(data.get("price")), 2),
+        quantity=data.get("quantity")
         )
+    db.session.add(new_prod)
+    db.session.commit()
+    return (
+        f"Product #{new_prod.id} added successfully.",
+        200
+    )
 
-@app.route("/store/<int:product_id>", methods=["PUT"])
+@app.route("/admin/store/<int:product_id>", methods=["PUT"])
+@login_required
 def update_product(product_id):
     try:
         data = request.json
 
         try:
             if db.session.get(Product, product_id) == None:
-                raise LookupError
+                raise FileNotFoundError
+            if current_user.__auth__() != True:
+                raise PermissionError
             
             err_list = []
+            err_string = ""
             
-            if type(data.get("name")) is not type("e") or type(None):
+            if type(data["name"]) != type("e") and type(data["name"]) != type(None):
                 err_list.append("name must be string")
-            if type(data.get("price")) is not type(0.0) or type(None) or type(0):
+            if type(data["price"]) != type(0) and type(data["price"]) != type(None) and type(data["price"]) != type(0.0):
                 err_list.append("price must be numeric")
-            if type(data.get("quantity")) is not type(0) or type(None):
+            if type(data["quantity"]) != type(0) and type(data["quantity"]) != type(None):
                 err_list.append("quantity must be an integer")
             
             if len(err_list) > 0:
-                raise ValueError
-
-        except LookupError:
+                for item in err_list:
+                    err_string += f"{item}"
+                    if item != err_list[-1]:
+                        err_string += ", "
+                raise AttributeError
+            
+        except AttributeError:
             return (
-                f"Product #{product_id} could not be found.",
-                404
-            )
-        
-        except ValueError:
-            return (
-                f"Invalid format for update attributes: {err_list}"
+                f"Invalid format for product attributes: {err_string}.",
+                400
             )
         
         old_prod = db.session.get(Product, product_id)
@@ -171,7 +207,7 @@ def update_product(product_id):
         new_prod = Product(
             id=product_id,
             name=data.get("name"),
-            price=data.get("price"),
+            price=round(float(data.get("price")), 2),
             quantity=data.get("quantity")
         )
 
@@ -183,7 +219,7 @@ def update_product(product_id):
             f"Product #{product_id} updated successfully.\nOld Product: {old_prod.to_json()}\nNew Product: {new_prod.to_json()}",
             200
         )
-    except:
+    except FileNotFoundError:
         return (
             f"Product #{product_id} could not be found.",
             404
@@ -215,7 +251,7 @@ def admin_store() -> str:
     except ValueError:
         return redirect(url_for("home"))
     
-    if user.__auth__() == True: 
+    if user.__auth__(): 
         return render_template("adminStore.html")
     return redirect(url_for("unauthorized"))
 
@@ -248,6 +284,80 @@ def get_products() -> str:
     products = Product.query.all()
     return jsonify([p.to_json() for p in products])
 
+@app.route("/orders", methods=["GET"])
+@login_required
+def retrieve_orders():
+    orders = current_user.orders
+    return jsonify([o.to_json() for o in orders])
+
+@app.route("/orders/<int:order_id>", methods=["GET"])
+@login_required
+def retrieve_order(order_id):
+    try:
+        orders = current_user.orders
+        
+        for order in orders:
+            if order.id == order_id:
+                return jsonify(order.to_json()), 200
+            
+        if db.session.get(Order, order_id) == None:
+            raise LookupError
+        else:
+            raise PermissionError
+    except LookupError:
+        return (
+            f"Order #{order_id} could not be found.",
+            404
+        )
+    except PermissionError:
+        return (
+            f"You do not have permission to view order #{order_id}.", #make this work with current unauthorized access code
+            401
+        )
+
+
+@app.route("/orders/neworder", methods=["POST"])
+@login_required
+def create_order():
+    data = request.json
+    new_order = Order(
+        user_id=current_user.id
+    )
+
+    db.session.add(new_order)
+    db.session.commit()
+    
+    for item in data["products"]:
+        line = OrderLine(
+            product_id=item["p_id"],
+            order_id=new_order.id,
+            quantity=item["quantity"]
+        )
+        db.session.add(line)
+        db.session.commit()
+        new_order.products.append(line)
+    
+    db.session.add(new_order)
+    db.session.commit()
+
+    return (
+        f"Order #{new_order.id} created successfully.\n{new_order.to_json()}",
+        200
+    )
+
+@app.route("/orders/<int:order_id>/<int:product_id>", methods=["DELETE"])
+@login_required
+def remove_order_line(order_id, product_id):
+    order = db.session.get(Order, order_id)
+
+    for product in order.products:
+        if product.product_id == product_id:
+            db.session.delete(product)
+            db.session.commit()
+            return (
+                f"Product #{product_id} successfully removed from order #{order_id}.",
+                200
+            )
 
 @app.route("/store/<int:id>")
 @login_required
